removed redundant mail template
