+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "xin",
-  "version": "0.8.2",
+  "version": "0.8.3",
   "description": "xin client",
   "main": "pclient-build",
   "scripts": {
